+++ src/engine.js
@@ -497,7 +497,19 @@ class Engine
             fragment: {
                 module: shader.getFragmentModule(),
                 targets: [{
-                    format: this.#format
+                    format: this.#format,
+                    blend: {
+                        color: {
+                            srcFactor: 'src-alpha',
+                            dstFactor: 'one-minus-src-alpha',
+                            operation: 'add'
+                        },
+                        alpha: {
+                            srcFactor: 'one',
+                            dstFactor: 'one-minus-src-alpha',
+                            operation: 'add'
+                        }
+                    }
                 }]
             },
             primitive: {