+++ changelog.md
@@ -3,6 +3,15 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Added
+
+- Add transparency support to rendering engine
+- Add Timer class for getting delta time
+- Add Getters for left, right, top, bottom, near and far in OrthographicCamera
+- Add Setters for left, right, top, bottom, near and far in OrthographicCamera
+
 ## [0.1.2] - 2026-01-01
 
 ### Fixed